Market place, Large refactoring (more to come), a couple of bug fixes
added the ability to toggle between the market place items available to
buy and the characters items available for sale...

Still need to add the ability to actually sell the items... although
most of the code is already there

+++ ACS/src/com/starfyre1/dataModel/MagicItemRecord.java
@@ -56,7 +56,7 @@ public class MagicItemRecord extends TKTableRecord {
 	public Object getRecord(int id) {
 		Object output = switch (id) {
 			case 0:
-				yield Float.valueOf(mCount);
+				yield Integer.valueOf(mCount);
 			case 1:
 				yield Boolean.valueOf(mEquipped);
 			case 2:
